Start on a async dns resolver
Since basic dns resolving functions like getaddrinfo() are synchronous,
using them in a latency-critical app requires threading support. The
simplest solution to this problem is to use a subprocess which will
do actual resolving, and a data pipe. The parent process could then
write a hostname to the request pipe and then epoll() on a response
pipe.